requirements.txt
numpy==2.0.2
pillow==11.0.0
tensorflow==2.18.0
fastapi==0.115.6